Move Album comparisons into the tool
- Remove tests for now (since they are not accessible to the test target from the site_tool)
